Add files via upload
This is a 'mini' report I made to demonstrate different examples of how Validation split and Dropout will effect a neural network. The code provided is what i used to create the graphics i have in the README file. The only differences were the validation split and dropout rate. One jupyter notebook goes through the process normally as explained but the 2nd notebook has a check point for the network.